产品管理方面全部

+++ src/com/shopping/controller/ProductImageController.java
@@ -13,6 +13,10 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.shopping.entity.Product;
 import com.shopping.entity.ProductImage;
 import com.shopping.service.ProductImageService;
+import com.shopping.util.CommonTools;
+import com.shopping.util.ImageUtil;
+import com.shopping.util.SnowFlake;
+import com.shopping.util.ValTool;
 
 @Controller
 @RequestMapping("/productImage")
@@ -34,26 +38,47 @@ public class ProductImageController {
 	
 	@RequestMapping("/deleteProductImage")
 	@ResponseBody
-	public Integer deleteProductImage(Integer productImageId) {
+	public Integer deleteProductImage(Integer productImageId,String imagePath) {
+		CommonTools.deleteFile(ValTool.PROJECT_PATH+imagePath);
+		if (CommonTools.deleteFile(ValTool.REAL_PATH+imagePath) == false) {
+			return 0;
+		}
 		return piService.delete(productImageId);
 	}
 	
 	@RequestMapping("/addProductImage")
 	@ResponseBody
-	public Integer addProductImage(File imageFile,BigInteger productId) {
+	public Integer addProductImage(BigInteger pId,String imageFile) {
 		
 		if (imageFile == null) {
-			return -1;
+			return 0;
+		}
+		
+		String relativePath = "";
+		if (imageFile != null && !(imageFile.equals(""))) {
+			ImageUtil imageUtil = new ImageUtil(imageFile);
+			
+			String contentPath = ValTool.REAL_PATH;
+			String suffix = imageUtil.getSuffix();
+			
+			System.out.println(contentPath);
+			String newFileName = SnowFlake.getId() + "." + suffix;
+			String rootPath = ValTool.CONTEXT_PATH;
+			relativePath = "/images/productImage/" + newFileName;
+			
+			System.out.println(ValTool.PROJECT_PATH+relativePath);
+			imageUtil.writeImage(ValTool.PROJECT_PATH+relativePath);
+			imageUtil.writeImage(contentPath+relativePath);
 		}
 		
 		Product product = new Product();
-		product.setProductId(productId);
+		product.setProductId(pId);
 		ProductImage productImage = new ProductImage();
 		productImage.setProduct(product);
-		productImage.setProductImagePath(imageFile.getPath());
+		productImage.setProductImagePath(relativePath);
 		
-		piService.addProductImageProductImage(productImage, productId);
+		Integer add = piService.addProductImageProductImage(productImage, pId);
 		
-		return 1;
+		return add;
 	}
 }
